requirements.txt
Django>=4.2,<5.0
django-admin-interface
django-crispy-forms==1.14.0
pillow
pymysql
mysqlclient
django-jazzmin
fontawesomefree
requests
openpyxl
django-rest-framework
django-paypal
paypal-checkout-serversdk
pytz
djangorestframework
pyodbc
sqlparse
asgiref
bcrypt
mssql-django
